+++ src/ducks/tooltips/types.js
@@ -1,0 +1,5 @@
+const tooltipsTypes = {
+  SET_ACTIVE_TOOLTIP: 'application/SET_ACTIVE_TOOLTIP',
+};
+
+export default tooltipsTypes;