+++ src/ducks/users/reducer.js
@@ -1,9 +1,11 @@
 import usersTypes from './types';
+import authTypes from '../auth/types';
 
 const {
   GET_ALL_USERS_SUCCESS,
   CREATE_USER_SUCCESS,
   DELETE_USER_SUCCESS,
+  CHANGE_USER_ROLE_SUCCESS,
 } = usersTypes;
 
 const initialState = {
@@ -11,7 +13,7 @@ const initialState = {
 };
 
 export default function reducer(state = initialState, action) {
-  const { type, data, error } = action;
+  const { type, data } = action;
   switch (type) {
     case CREATE_USER_SUCCESS:
       return {
@@ -23,11 +25,29 @@ export default function reducer(state = initialState, action) {
         ...state,
         list: state.list.filter(e => e.id !== data.id),
       };
+    case CHANGE_USER_ROLE_SUCCESS: {
+      const updatedList = state.list.map(p =>
+        p.id === data.id ? { ...p, role: data.role } : p,
+      );
+      return {
+        ...state,
+        list: updatedList,
+      };
+    }
     case GET_ALL_USERS_SUCCESS:
       return {
         ...state,
         list: data,
       };
+    case authTypes.login.FORGOT_PASSWORD_SUCCESS: {
+      const updatedList = state.list.map(p =>
+        p.id === data.id ? { ...p, redirectUrl: data.redirectUrl } : p,
+      );
+      return {
+        ...state,
+        list: updatedList,
+      };
+    }
     default:
       return state;
   }