+++ src/ducks/users/sagas.js
@@ -9,8 +9,17 @@ import { applicationStates } from 'types/applicationStates';
 function* createUserSaga({ data }) {
   try {
     yield put(applicationActions.updateStatus(applicationStates.BUSY));
-    yield call(usersService.createUser, data);
-    yield put(usersActions.createUserSuccess(data));
+    const res = yield call(usersService.createUser, data);
+    // eslint-disable-next-line camelcase
+    const { id, registration_url } = res.data;
+    delete data.redirect_url;
+    yield put(
+      usersActions.createUserSuccess({
+        ...data,
+        id,
+        redirectUrl: registration_url,
+      }),
+    );
     yield put(applicationActions.notification({ title: `User added` }));
     yield put(applicationActions.updateStatus(applicationStates.IDLE));
   } catch (error) {
@@ -41,8 +50,20 @@ function* getAllUsersSaga() {
   }
 }
 
+function* changeUserRoleSaga({ data }) {
+  try {
+    yield put(applicationActions.updateStatus(applicationStates.BUSY));
+    yield call(usersService.changeUserRole, data);
+    yield put(usersActions.changeUserRoleSuccess(data));
+    yield put(applicationActions.updateStatus(applicationStates.IDLE));
+  } catch (error) {
+    yield call(errorHandlerSaga, error);
+  }
+}
+
 export function* usersSaga() {
   yield takeEvery(usersTypes.CREATE_USER_REQUEST, createUserSaga);
   yield takeEvery(usersTypes.DELETE_USER_REQUEST, deleteUser);
   yield takeEvery(usersTypes.GET_ALL_USERS_REQUEST, getAllUsersSaga);
+  yield takeEvery(usersTypes.CHANGE_USER_ROLE_REQUEST, changeUserRoleSaga);
 }