+++ src/views/Settings/Members/Dropdown/index.js
@@ -11,14 +11,27 @@ import {
 import { useOnClickOutside } from '../../../../hooks/useOnClickOutside';
 import { useDispatch } from 'react-redux';
 import usersActions from '../../../../ducks/users/actions';
+import authActions from '../../../../ducks/auth/actions';
+import registrationActions from '../../../../ducks/registration/actions';
 
-const Dropdown = ({ id }) => {
+const Dropdown = ({
+  id,
+  role: userRole,
+  scrollToItem,
+  setShowModal,
+  setPasswordReset,
+}) => {
   const containerRef = useRef();
   const dispatch = useDispatch();
   const [openMenu, setOpenMenu] = useState(false);
-  const [role, setRole] = useState('admin');
+  const [role, setRole] = useState(userRole);
 
   useOnClickOutside(containerRef, () => setOpenMenu(false));
+
+  const openMenuAndScroll = () => {
+    setOpenMenu(!openMenu);
+    scrollToItem();
+  };
 
   const deleteMember = () => {
     const isTheySure = window.confirm('Are you sure?');
@@ -30,8 +43,27 @@ const Dropdown = ({ id }) => {
   const changeTo = to => {
     const isTheySure = window.confirm('Are you sure?');
     if (isTheySure) {
-      // here we'll dispatch the action and the role will come via props
+      dispatch(usersActions.changeUserRoleRequest({ id, role: to }));
       setRole(to);
+    }
+  };
+
+  const resetPassword = () => {
+    if (id) {
+      dispatch(
+        authActions.forgotPassword({
+          id,
+          redirect_url: `${process.env.REACT_APP_BASE_URL}reset-password`,
+        }),
+      );
+      setPasswordReset(id);
+      setShowModal(true);
+    }
+  };
+
+  const resetPhoneNumber = () => {
+    if (id) {
+      dispatch(registrationActions.removeMfa({ id }));
     }
   };
 
@@ -40,31 +72,31 @@ const Dropdown = ({ id }) => {
       case 'admin': {
         return (
           <>
-            <div onClick={() => changeTo('superadmin')} className={item}>
+            <div onClick={() => changeTo('super_admin')} className={item}>
               Change to Super Admin
             </div>
-            <div onClick={() => changeTo('tracer')} className={item}>
+            <div onClick={() => changeTo('contact_tracer')} className={item}>
               Change to Contact Tracer
             </div>
           </>
         );
       }
-      case 'superadmin': {
+      case 'super_admin': {
         return (
           <>
             <div onClick={() => changeTo('admin')} className={item}>
               Change to Admin
             </div>
-            <div onClick={() => changeTo('tracer')} className={item}>
+            <div onClick={() => changeTo('contact_tracer')} className={item}>
               Change to Contact Tracer
             </div>
           </>
         );
       }
-      case 'tracer': {
+      case 'contact_tracer': {
         return (
           <>
-            <div onClick={() => changeTo('superadmin')} className={item}>
+            <div onClick={() => changeTo('super_admin')} className={item}>
               Change to Super Admin
             </div>
             <div onClick={() => changeTo('admin')} className={item}>
@@ -78,15 +110,17 @@ const Dropdown = ({ id }) => {
     }
   };
   return (
-    <div
-      onClick={() => setOpenMenu(!openMenu)}
-      className={dropdown}
-      ref={containerRef}
-    >
+    <div onClick={openMenuAndScroll} className={dropdown} ref={containerRef}>
       <FontAwesomeIcon icon={faEllipsisV} className={menuIcon} />
       {openMenu && (
         <div id="dropdownMenu" className={dropdownContent}>
           {renderChangeTo()}
+          <div onClick={resetPassword} className={item}>
+            Reset password
+          </div>
+          <div onClick={resetPhoneNumber} className={item}>
+            Reset phone number
+          </div>
           <div onClick={deleteMember} className={`${item} ${deleteMemberText}`}>
             Delete member
           </div>