+++ changelog.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## v2.3.0 - Tuesday, August 18, 2020
+
+### Updates
+* Admin members CRUD completion
+* User account creation flow (Pending MFA full flow due SMS service)
+* Hide sections depending on user role
+* UI Fixes and changes
+* Bug Fixes
+
+## v2.2.0-alpha - Monday, August 3, 2020
+
+### Updates
+* Add reset password flow
+* Improve UI transitions
+* Various minor bug fixes
+* Get all users for a specific HA
+* Create user
+* Delete User
+
 ## v2.1.0-alpha - Tuesday, July 28, 2020
 
 ### Updates